# api_documentation/API-Documentation-ASE23-GREEN-resolved.yaml
...
openapi: 3.0.1
info:
  title: ASE23-GREEN
  description: "API documentation for Université de Fribourg (Switzerland) ASE2023 Project - Group Green. All API calls except login and register need an authorization token in the request. Simply add an 'authorization' header within the request. Testing the API documentation directly on SwaggerHub does not work. Use the Swaggerhub Explorer or another program such as Postman to perform the API calls. We are unsure about the cause of this issue. The localhost can be used when the API is running on the local machine, whereas the 'https://soizhiwen.com/api' website can be used to call the API via our hosted server which also enables to ineract with the accompanying interface reachable via 'https://soizhiwen.com'. This server is accessible at the time of writing this, December 2023."
  version: "0.93"
servers:
- url: https://soizhiwen.com/api
- url: http://localhost:8080
paths:
...
  /register:
    post:
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/register_body'
            examples:
              example:
                value: "{\r\n    \"name\": \"Admin1\",\r\n    \"email\": \"Admin1@mail.com\",\r\n    \"password\": \"Admin1PW\",\r\n    \"role\": \"Admin\"\r\n}"
      responses:
        "200":
          description: Creates a new Admin. No authorization token needed. The response contains the respective authorization token in the 'authorization' header.
  /login:
    post:
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/login_body'
            examples:
              example:
                value: "{\r\n    \"email\": \"Admin1@mail.com\",\r\n    \"password\": \"Admin1PW\"\r\n}"
      responses:
        "200":
          description: Logs-in an existing Admin or User. No authorization token needed. The response contains the respective authorization token in the 'authorization' header.
...
components:
  schemas:
    register_body:
      required:
      - email
      - name
      - password
      - role
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
      additionalProperties: false
    login_body:
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      additionalProperties: false
...
  securitySchemes: {}